message

+++ MyOrthoClient/MyOrthoClient/Models/ActivityVM.cs
@@ -34,7 +34,7 @@ namespace MyOrthoClient.Models
         public bool Courbe_f0_exacteEvaluated { get; set; }
         public int Duree_exacte { get; set; }
         public bool Duree_exacteEvaluated { get; set; }
-        public int Jitter { get; set; }
+        public double Jitter { get; set; }
         public bool JitterEvaluated { get; set; }
         public ICollection<DataLineItem> Exercice
         {
